[libc++] Start tracking Github issues in status pages (#149833)
This patch adds another row to the Status pages that cross-references
the Github issue. It also ensures that the synchronization script takes
that new row into account.

This should make it easier to find out about the detailed status of a
paper from the status pages by clicking on the link and being taken
directly to its associated Github issue. I expect that this should
remove the need for many "Notes" which simply duplicate the information
of which parts of a paper are implemented: instead we can list that the
implementation is partial and users can click on the Github issue to see
what's implemented.

NOKEYCHECK=True
GitOrigin-RevId: e0d0b90d52f54f291c1042046ddf22e925408235