[libc++] Disable atomic_wait benchmarks outside of dry-run mode (#158631)
The atomic_wait benchmarks are great, but they tend to overload the
system they're running on. For that reason, we can't run them on our CI
infrastructure on a regular basis.

Instead of removing them, make them unsupported outside of dry-running,
which allows keeping the benchmarks around and ensuring they don't rot,
but doesn't run them along with the other benchmarks. If we need to
investigate atomic_wait performance, it's trivial to mark the benchmark
as supported and run it for local investigations.

This is an alternative to https://github.com/llvm/llvm-project/pull/158289.

NOKEYCHECK=True
GitOrigin-RevId: 815b1641a65c33dd331efb816fdafeafd0a7e794