[libc++] Improve handling of runtime errors inside SPEC benchmarks
Previously, we would report a successful run if the benchmark exited
with an error, and we would produce a timing for the benchmark. After
this patch, we consider an error in the benchmark to be a failed LIT
test and we don't produce any benchmark data for it.

NOKEYCHECK=True
GitOrigin-RevId: 62450ba905fecc0821f3a08db81f756c8b1f7b7c