[libc++][ranges] Fix `ranges::join_view` segmented iterator trait (#158347)
The outer iterator needs to move to the next segment when calling
__compose.

Without this change, `find_segment_if` would never reach the end of the
join_view which caused erroneous result when calling `ranges::find` on a
join_view of bidirectional ranges.

Other specializations using the segmented iterator trait were likely to
be affected as well.

Fixes #158279
Fixes #93180

NOKEYCHECK=True
GitOrigin-RevId: d1b5607dc113016b74d0a58e95fed00ea9ad7950