[libcxx] Run tests on Windows/arm64 too (#161906)
Github Actions has got free runners with Windows on ARM64 these days,
which can be used for running CI, just as well as the other existing
cases.

This qualifies this configuration as a supported target platform, thus
add it to the docs in the listing of supported platforms.

NOKEYCHECK=True
GitOrigin-RevId: 646bc09d17de317c139c9e49af308ab1ea2c9641