[libc++] Implement our own is{,x}digit functions for the C locale (#165467)
The C locale is defined by the C standard, so we know exactly which
digits classify as (x)digits. Instead of going through the locale base
API we can simply implement functions which determine whether a
character is one ourselves, and probably improve codegen significantly
as well that way.

NOKEYCHECK=True
GitOrigin-RevId: 693f700e1dba5d601d3ac999ff5471a3d0aaac70