Revert "[libc++] Fix the locale base API on Linux with musl" (#169884)
The patch has been committed without approval from the libc++ review
group and is implementing the locale base API in a way it wasn't
intended to be implemented. The commit also contains a no-reply github
E-Mail, which is in conflict with the LLVM developer policy.

Reverts llvm/llvm-project#167980

NOKEYCHECK=True
GitOrigin-RevId: 26b82e5b8248335ae7b9db76a7da9dd0587d5baa